Add commandline argument parser
It is often useful to be able to directly specify a serial port and its
settings from the commandline rather than having to select it in the GUI
each time.

Add a CLI parser based on `gumdrop` to parse commandline arguments.  The
selected serial port is then immediately opened on startup.